chunk hometab into components cus it was too large
