+++ app/src/main/java/com/bundl/app/presentation/dashboard/components/EmptyOrdersCardComponent.kt
@@ -1,0 +1,54 @@
+package com.bundl.app.presentation.dashboard.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.bundl.app.presentation.theme.BundlColors
+
+@Composable
+fun EmptyOrdersCard() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center,
+            modifier = Modifier
+                .fillMaxSize()
+                .background(
+                    color = BundlColors.SurfaceLight,
+                    shape = RoundedCornerShape(8.dp)
+                )
+                .padding(24.dp)
+        ) {
+            Icon(
+                imageVector = Icons.Default.Info,
+                contentDescription = "Info",
+                tint = Color.LightGray,
+                modifier = Modifier
+                    .size(48.dp)
+                    .padding(bottom = 16.dp)
+            )
+            Text(
+                text = "No Active Orders Available",
+                style = MaterialTheme.typography.titleMedium,
+                fontWeight = FontWeight.Bold,
+                color = Color.White,
+                textAlign = TextAlign.Center
+            )
+        }
+    }
+}
